+++ Changes.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+
+- Add `pre_process` block support
+
 0.5.0
 -----